requirements.txt
Flask==3.0.2
tensorflow==2.15.0
numpy<2.0          # Forces installation of a compatible NumPy 1.x version
gunicorn==20.1.0
Pillow==9.4.0